requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
sqlalchemy>=1.4.0
pyyaml>=6.0

# Excel processing
openpyxl>=3.0.9
xlrd>=2.0.1

# Database
sqlite3

# Date/time handling
python-dateutil>=2.8.0

# Configuration
pydantic>=1.10.0

# CLI and utilities
click>=8.0.0
rich>=12.0.0
tabulate>=0.9.0

# Performance monitoring
psutil>=5.9.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
mutmut>=3.3.0

# Pre-commit hooks and code quality
pre-commit>=3.0.0
isort>=5.13.0
pylint>=3.0.0
bandit>=1.7.5

# Future ML dependencies (for the smart recommendation system discussed)
# scikit-learn>=1.1.0
# spacy>=3.4.0
# transformers>=4.21.0
# sentence-transformers>=2.2.0
# mlflow>=1.28.0
# optuna>=3.0.0
